Quick madvise hook fix: remove premature return in hooks (#134)
* madvise probe properly handle do_madvise call result and do_vmi_align_munmap call

madvise probes were not properly collecting data

* collecting raw numerical data with correct unsigned types

+++ python/kernmlops/data_collection/bpf_instrumentation/madvise.py
@@ -38,10 +38,10 @@ ADVICE_ASSIGN_DICT = {
 
 @dataclass(frozen=True)
 class MadviseStat:
-  tgid: int
-  ts_ns: int
-  address: int
-  length: int
+  tgid: pl.UInt32
+  ts_ns: pl.UInt64
+  address: pl.UInt64
+  length: pl.UInt64
   advice: str
 
 class MadviseBPFHook(BPFProgram):
